+++ src/pages/classes/ClassModal/Steps/First/class.modal.steps.first.interface.ts
@@ -1,0 +1,4 @@
+export interface ClassModalStepsProps {
+  title: string;
+  description: string;
+}
